operations/scripts/requirements.txt
attrs==22.2.0
boto3==1.26.133
botocore==1.29.133
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
contextlib2==21.6.0
dill==0.3.6
google-pasta==0.2.0
graphviz==0.8.4
idna==3.4
imageio==2.28.1
importlib-metadata==4.13.0
importlib-resources==5.12.0
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
lazy_loader==0.2
multiprocess==0.70.14
mxnet==1.9.1
mxnet-mkl==1.6.0
networkx==3.1
numpy==1.23.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pathos==0.3.0
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pox==0.3.2
ppft==1.7.6.6
protobuf==3.20.3
protobuf3-to-dict==0.1.5
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
requests==2.30.0
s3transfer==0.6.1
sagemaker==2.154.0
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
smdebug-rulesconfig==1.0.1
tblib==1.7.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
